Fixes that invalid songs could sometimes not be removed from playlists
During sorting a playlist, invalid song GUIDs got cleared to empty strings. This resulted in a playlist that had a proper (invalid) GUID in the original data, but when trying to delete it, the GUID was corrupted (empty). Thus the deletion did not reference the proper item and did not persist across runs.